osbuild: xattr implementation
An initial implementation of extended attributes in the utility module,
this uses libc to, specifically on Linux: list, remove, set, or get,
extended attributes and their values.

It also provides an extension trait on `cap_std::fs::File` for some
extra methods there to work with xattrs.

This is preliminary work for `copy_tree`.